requirements.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.4.0
django-shortuuidfield==0.1.3
djangorestframework==3.11.1
PyJWT==1.7.1
pytz==2020.1
shortuuid==1.0.1
six==1.15.0
sqlparse==0.3.1
